Move DDNet settings to `menus_settings_ddnet.cpp`
