+++ src/game/client/components/menus_settings_ddnet.cpp
@@ -1,0 +1,424 @@
+/* (c) Magnus Auvinen. See licence.txt in the root of the distribution for more information. */
+/* If you are missing that file, acquire a complete release at teeworlds.com.                */
+#include "menus.h"
+
+#include <base/dbg.h>
+#include <base/fs.h>
+#include <base/str.h>
+
+#include <engine/font_icons.h>
+#include <engine/graphics.h>
+#include <engine/shared/config.h>
+#include <engine/storage.h>
+#include <engine/textrender.h>
+#include <engine/updater.h>
+
+#include <game/client/components/background.h>
+#include <game/client/components/mapimages.h>
+#include <game/client/components/tooltips.h>
+#include <game/client/gameclient.h>
+#include <game/client/ui.h>
+#include <game/client/ui_listbox.h>
+#include <game/localization.h>
+
+#include <algorithm>
+
+void CMenus::RenderSettingsDDNet(CUIRect MainView)
+{
+	CUIRect Button, Left, Right, LeftLeft, Label;
+
+#if defined(CONF_AUTOUPDATE)
+	CUIRect UpdaterRect;
+	MainView.HSplitBottom(20.0f, &MainView, &UpdaterRect);
+	MainView.HSplitBottom(5.0f, &MainView, nullptr);
+#endif
+
+	// demo
+	CUIRect Demo;
+	MainView.HSplitTop(110.0f, &Demo, &MainView);
+	Demo.HSplitTop(30.0f, &Label, &Demo);
+	Ui()->DoLabel(&Label, Localize("Demo"), 20.0f, TEXTALIGN_ML);
+	Label.VSplitMid(nullptr, &Label, 20.0f);
+	Ui()->DoLabel(&Label, Localize("Ghost"), 20.0f, TEXTALIGN_ML);
+
+	Demo.HSplitTop(5.0f, nullptr, &Demo);
+	Demo.VSplitMid(&Left, &Right, 20.0f);
+
+	Left.HSplitTop(20.0f, &Button, &Left);
+	if(DoButton_CheckBox(&g_Config.m_ClAutoRaceRecord, Localize("Save the best demo of each race"), g_Config.m_ClAutoRaceRecord, &Button))
+	{
+		g_Config.m_ClAutoRaceRecord ^= 1;
+	}
+
+	Left.HSplitTop(20.0f, &Button, &Left);
+	if(DoButton_CheckBox(&g_Config.m_ClReplays, Localize("Enable replays"), g_Config.m_ClReplays, &Button))
+	{
+		g_Config.m_ClReplays ^= 1;
+		if(Client()->State() == IClient::STATE_ONLINE)
+		{
+			Client()->DemoRecorder_UpdateReplayRecorder();
+		}
+	}
+
+	Left.HSplitTop(20.0f, &Button, &Left);
+	if(g_Config.m_ClReplays)
+		Ui()->DoScrollbarOption(&g_Config.m_ClReplayLength, &g_Config.m_ClReplayLength, &Button, Localize("Default length"), 10, 600, &CUi::ms_LinearScrollbarScale, CUi::SCROLLBAR_OPTION_NOCLAMPVALUE);
+
+	Right.HSplitTop(20.0f, &Button, &Right);
+	if(DoButton_CheckBox(&g_Config.m_ClRaceGhost, Localize("Enable ghost"), g_Config.m_ClRaceGhost, &Button))
+	{
+		g_Config.m_ClRaceGhost ^= 1;
+	}
+	GameClient()->m_Tooltips.DoToolTip(&g_Config.m_ClRaceGhost, &Button, Localize("When you cross the start line, show a ghost tee replicating the movements of your best time"));
+
+	if(g_Config.m_ClRaceGhost)
+	{
+		Right.HSplitTop(20.0f, &Button, &Right);
+		Button.VSplitMid(&LeftLeft, &Button);
+		if(DoButton_CheckBox(&g_Config.m_ClRaceShowGhost, Localize("Show ghost"), g_Config.m_ClRaceShowGhost, &LeftLeft))
+		{
+			g_Config.m_ClRaceShowGhost ^= 1;
+		}
+		Ui()->DoScrollbarOption(&g_Config.m_ClRaceGhostAlpha, &g_Config.m_ClRaceGhostAlpha, &Button, Localize("Opacity"), 0, 100, &CUi::ms_LinearScrollbarScale, 0u, "%");
+
+		Right.HSplitTop(20.0f, &Button, &Right);
+		if(DoButton_CheckBox(&g_Config.m_ClRaceSaveGhost, Localize("Save ghost"), g_Config.m_ClRaceSaveGhost, &Button))
+		{
+			g_Config.m_ClRaceSaveGhost ^= 1;
+		}
+
+		if(g_Config.m_ClRaceSaveGhost)
+		{
+			Right.HSplitTop(20.0f, &Button, &Right);
+			if(DoButton_CheckBox(&g_Config.m_ClRaceGhostSaveBest, Localize("Only save improvements"), g_Config.m_ClRaceGhostSaveBest, &Button))
+			{
+				g_Config.m_ClRaceGhostSaveBest ^= 1;
+			}
+		}
+	}
+
+	// gameplay
+	CUIRect Gameplay;
+	MainView.HSplitTop(170.0f, &Gameplay, &MainView);
+	Gameplay.HSplitTop(30.0f, &Label, &Gameplay);
+	Ui()->DoLabel(&Label, Localize("Gameplay"), 20.0f, TEXTALIGN_ML);
+	Gameplay.HSplitTop(5.0f, nullptr, &Gameplay);
+	Gameplay.VSplitMid(&Left, &Right, 20.0f);
+
+	Left.HSplitTop(20.0f, &Button, &Left);
+	Ui()->DoScrollbarOption(&g_Config.m_ClOverlayEntities, &g_Config.m_ClOverlayEntities, &Button, Localize("Overlay entities"), 0, 100);
+
+	Left.HSplitTop(20.0f, &Button, &Left);
+	Button.VSplitMid(&LeftLeft, &Button);
+
+	if(DoButton_CheckBox(&g_Config.m_ClTextEntities, Localize("Show text entities"), g_Config.m_ClTextEntities, &LeftLeft))
+		g_Config.m_ClTextEntities ^= 1;
+
+	if(g_Config.m_ClTextEntities)
+	{
+		if(Ui()->DoScrollbarOption(&g_Config.m_ClTextEntitiesSize, &g_Config.m_ClTextEntitiesSize, &Button, Localize("Size"), 20, 100, &CUi::ms_LinearScrollbarScale, CUi::SCROLLBAR_OPTION_DELAYUPDATE))
+			GameClient()->m_MapImages.SetTextureScale(g_Config.m_ClTextEntitiesSize);
+	}
+
+	Left.HSplitTop(20.0f, &Button, &Left);
+	Button.VSplitMid(&LeftLeft, &Button);
+
+	if(DoButton_CheckBox(&g_Config.m_ClShowOthers, Localize("Show others"), g_Config.m_ClShowOthers == SHOW_OTHERS_ON, &LeftLeft))
+		g_Config.m_ClShowOthers = g_Config.m_ClShowOthers != SHOW_OTHERS_ON ? SHOW_OTHERS_ON : SHOW_OTHERS_OFF;
+
+	Ui()->DoScrollbarOption(&g_Config.m_ClShowOthersAlpha, &g_Config.m_ClShowOthersAlpha, &Button, Localize("Opacity"), 0, 100, &CUi::ms_LinearScrollbarScale, 0u, "%");
+
+	GameClient()->m_Tooltips.DoToolTip(&g_Config.m_ClShowOthersAlpha, &Button, Localize("Adjust the opacity of entities belonging to other teams, such as tees and name plates"));
+
+	Left.HSplitTop(20.0f, &Button, &Left);
+	static int s_ShowOwnTeamId = 0;
+	if(DoButton_CheckBox(&s_ShowOwnTeamId, Localize("Show others (own team only)"), g_Config.m_ClShowOthers == SHOW_OTHERS_ONLY_TEAM, &Button))
+	{
+		g_Config.m_ClShowOthers = g_Config.m_ClShowOthers != SHOW_OTHERS_ONLY_TEAM ? SHOW_OTHERS_ONLY_TEAM : SHOW_OTHERS_OFF;
+	}
+
+	Left.HSplitTop(20.0f, &Button, &Left);
+	if(DoButton_CheckBox(&g_Config.m_ClShowQuads, Localize("Show background quads"), g_Config.m_ClShowQuads, &Button))
+	{
+		g_Config.m_ClShowQuads ^= 1;
+	}
+	GameClient()->m_Tooltips.DoToolTip(&g_Config.m_ClShowQuads, &Button, Localize("Quads are used for background decoration"));
+
+	Left.HSplitTop(20.0f, &Button, &Left);
+	if(Ui()->DoScrollbarOption(&g_Config.m_ClDefaultZoom, &g_Config.m_ClDefaultZoom, &Button, Localize("Default zoom"), 0, 20))
+		GameClient()->m_Camera.SetZoom(CCamera::ZoomStepsToValue(g_Config.m_ClDefaultZoom - 10), g_Config.m_ClSmoothZoomTime, true);
+
+	Right.HSplitTop(20.0f, &Button, &Right);
+	Ui()->DoScrollbarOption(&g_Config.m_ClPredictionMargin, &g_Config.m_ClPredictionMargin, &Button, Localize("Prediction margin"), 1, 300);
+
+	Right.HSplitTop(20.0f, &Button, &Right);
+	if(DoButton_CheckBox(&g_Config.m_ClPredictEvents, Localize("Predict events (experimental)"), g_Config.m_ClPredictEvents, &Button))
+	{
+		g_Config.m_ClPredictEvents ^= 1;
+	}
+
+	Right.HSplitTop(20.0f, &Button, &Right);
+	if(DoButton_CheckBox(&g_Config.m_ClAntiPing, Localize("AntiPing"), g_Config.m_ClAntiPing, &Button))
+	{
+		g_Config.m_ClAntiPing ^= 1;
+	}
+	GameClient()->m_Tooltips.DoToolTip(&g_Config.m_ClAntiPing, &Button, Localize("Tries to predict other entities to give a feel of low latency"));
+
+	if(g_Config.m_ClAntiPing)
+	{
+		Right.HSplitTop(20.0f, &Button, &Right);
+		if(DoButton_CheckBox(&g_Config.m_ClAntiPingPlayers, Localize("AntiPing: predict other players"), g_Config.m_ClAntiPingPlayers, &Button))
+		{
+			g_Config.m_ClAntiPingPlayers ^= 1;
+		}
+
+		Right.HSplitTop(20.0f, &Button, &Right);
+		if(DoButton_CheckBox(&g_Config.m_ClAntiPingWeapons, Localize("AntiPing: predict weapons"), g_Config.m_ClAntiPingWeapons, &Button))
+		{
+			g_Config.m_ClAntiPingWeapons ^= 1;
+		}
+
+		Right.HSplitTop(20.0f, &Button, &Right);
+		if(DoButton_CheckBox(&g_Config.m_ClAntiPingGrenade, Localize("AntiPing: predict grenade paths"), g_Config.m_ClAntiPingGrenade, &Button))
+		{
+			g_Config.m_ClAntiPingGrenade ^= 1;
+		}
+	}
+
+	CUIRect Background, Miscellaneous;
+	MainView.VSplitMid(&Background, &Miscellaneous, 20.0f);
+
+	// background
+	Background.HSplitTop(30.0f, &Label, &Background);
+	Background.HSplitTop(5.0f, nullptr, &Background);
+	Ui()->DoLabel(&Label, Localize("Background"), 20.0f, TEXTALIGN_ML);
+
+	ColorRGBA GreyDefault(0.5f, 0.5f, 0.5f, 1);
+
+	static CButtonContainer s_ResetId1;
+	DoLine_ColorPicker(&s_ResetId1, 25.0f, 13.0f, 5.0f, &Background, Localize("Regular background color"), &g_Config.m_ClBackgroundColor, GreyDefault, false);
+
+	static CButtonContainer s_ResetId2;
+	DoLine_ColorPicker(&s_ResetId2, 25.0f, 13.0f, 5.0f, &Background, Localize("Entities background color"), &g_Config.m_ClBackgroundEntitiesColor, GreyDefault, false);
+
+	CUIRect EditBox, ReloadButton;
+	Background.HSplitTop(20.0f, &Label, &Background);
+	Background.HSplitTop(2.0f, nullptr, &Background);
+	Label.VSplitLeft(100.0f, &Label, &EditBox);
+	EditBox.VSplitRight(60.0f, &EditBox, &Button);
+	Button.VSplitMid(&ReloadButton, &Button, 5.0f);
+	EditBox.VSplitRight(5.0f, &EditBox, nullptr);
+
+	Ui()->DoLabel(&Label, Localize("Map"), 14.0f, TEXTALIGN_ML);
+
+	static CLineInput s_BackgroundEntitiesInput(g_Config.m_ClBackgroundEntities, sizeof(g_Config.m_ClBackgroundEntities));
+	Ui()->DoEditBox(&s_BackgroundEntitiesInput, &EditBox, 14.0f);
+
+	static CButtonContainer s_BackgroundEntitiesMapPicker, s_BackgroundEntitiesReload;
+
+	if(Ui()->DoButton_FontIcon(&s_BackgroundEntitiesReload, FontIcon::ARROW_ROTATE_RIGHT, 0, &ReloadButton, BUTTONFLAG_LEFT))
+	{
+		GameClient()->m_Background.LoadBackground();
+	}
+
+	if(Ui()->DoButton_FontIcon(&s_BackgroundEntitiesMapPicker, FontIcon::FOLDER, 0, &Button, BUTTONFLAG_LEFT))
+	{
+		static SPopupMenuId s_PopupMapPickerId;
+		static CPopupMapPickerContext s_PopupMapPickerContext;
+		s_PopupMapPickerContext.m_pMenus = this;
+		s_PopupMapPickerContext.MapListPopulate();
+		Ui()->DoPopupMenu(&s_PopupMapPickerId, Ui()->MouseX(), Ui()->MouseY(), 300.0f, 250.0f, &s_PopupMapPickerContext, PopupMapPicker);
+	}
+
+	Background.HSplitTop(20.0f, &Button, &Background);
+	const bool UseCurrentMap = str_comp(g_Config.m_ClBackgroundEntities, CURRENT_MAP) == 0;
+	static int s_UseCurrentMapId = 0;
+	if(DoButton_CheckBox(&s_UseCurrentMapId, Localize("Use current map as background"), UseCurrentMap, &Button))
+	{
+		if(UseCurrentMap)
+			g_Config.m_ClBackgroundEntities[0] = '\0';
+		else
+			str_copy(g_Config.m_ClBackgroundEntities, CURRENT_MAP);
+		GameClient()->m_Background.LoadBackground();
+	}
+
+	Background.HSplitTop(20.0f, &Button, &Background);
+	if(DoButton_CheckBox(&g_Config.m_ClBackgroundShowTilesLayers, Localize("Show tiles layers from BG map"), g_Config.m_ClBackgroundShowTilesLayers, &Button))
+		g_Config.m_ClBackgroundShowTilesLayers ^= 1;
+
+	// miscellaneous
+	Miscellaneous.HSplitTop(30.0f, &Label, &Miscellaneous);
+	Miscellaneous.HSplitTop(5.0f, nullptr, &Miscellaneous);
+
+	Ui()->DoLabel(&Label, Localize("Miscellaneous"), 20.0f, TEXTALIGN_ML);
+
+	static CButtonContainer s_ButtonTimeout;
+	Miscellaneous.HSplitTop(20.0f, &Button, &Miscellaneous);
+	if(DoButton_Menu(&s_ButtonTimeout, Localize("New random timeout code"), 0, &Button))
+	{
+		Client()->GenerateTimeoutSeed();
+	}
+
+	Miscellaneous.HSplitTop(5.0f, nullptr, &Miscellaneous);
+	Miscellaneous.HSplitTop(20.0f, &Label, &Miscellaneous);
+	Miscellaneous.HSplitTop(2.0f, nullptr, &Miscellaneous);
+	Ui()->DoLabel(&Label, Localize("Run on join"), 14.0f, TEXTALIGN_ML);
+	Miscellaneous.HSplitTop(20.0f, &Button, &Miscellaneous);
+	static CLineInput s_RunOnJoinInput(g_Config.m_ClRunOnJoin, sizeof(g_Config.m_ClRunOnJoin));
+	s_RunOnJoinInput.SetEmptyText(Localize("Chat command (e.g. showall 1)"));
+	Ui()->DoEditBox(&s_RunOnJoinInput, &Button, 14.0f);
+
+#if defined(CONF_FAMILY_WINDOWS)
+	static CButtonContainer s_ButtonUnregisterShell;
+	Miscellaneous.HSplitTop(10.0f, nullptr, &Miscellaneous);
+	Miscellaneous.HSplitTop(20.0f, &Button, &Miscellaneous);
+	if(DoButton_Menu(&s_ButtonUnregisterShell, Localize("Unregister protocol and file extensions"), 0, &Button))
+	{
+		Client()->ShellUnregister();
+	}
+#endif
+
+	// Updater
+#if defined(CONF_AUTOUPDATE)
+	{
+		bool NeedUpdate = str_comp(Client()->LatestVersion(), "0");
+		IUpdater::EUpdaterState State = Updater()->GetCurrentState();
+
+		// Update Button
+		char aBuf[256];
+		if(NeedUpdate && State <= IUpdater::CLEAN)
+		{
+			str_format(aBuf, sizeof(aBuf), Localize("DDNet %s is available:"), Client()->LatestVersion());
+			UpdaterRect.VSplitLeft(TextRender()->TextWidth(14.0f, aBuf, -1, -1.0f) + 10.0f, &UpdaterRect, &Button);
+			Button.VSplitLeft(100.0f, &Button, nullptr);
+			static CButtonContainer s_ButtonUpdate;
+			if(DoButton_Menu(&s_ButtonUpdate, Localize("Update now"), 0, &Button))
+			{
+				Updater()->InitiateUpdate();
+			}
+		}
+		else if(State >= IUpdater::GETTING_MANIFEST && State < IUpdater::NEED_RESTART)
+			str_copy(aBuf, Localize("Updating…"));
+		else if(State == IUpdater::NEED_RESTART)
+		{
+			str_copy(aBuf, Localize("DDNet Client updated!"));
+			m_NeedRestartUpdate = true;
+		}
+		else
+		{
+			str_copy(aBuf, Localize("No updates available"));
+			UpdaterRect.VSplitLeft(TextRender()->TextWidth(14.0f, aBuf, -1, -1.0f) + 10.0f, &UpdaterRect, &Button);
+			Button.VSplitLeft(100.0f, &Button, nullptr);
+			static CButtonContainer s_ButtonUpdate;
+			if(DoButton_Menu(&s_ButtonUpdate, Localize("Check now"), 0, &Button))
+			{
+				Client()->RequestDDNetInfo();
+			}
+		}
+		Ui()->DoLabel(&UpdaterRect, aBuf, 14.0f, TEXTALIGN_ML);
+		TextRender()->TextColor(1.0f, 1.0f, 1.0f, 1.0f);
+	}
+#endif
+}
+
+CUi::EPopupMenuFunctionResult CMenus::PopupMapPicker(void *pContext, CUIRect View, bool Active)
+{
+	CPopupMapPickerContext *pPopupContext = static_cast<CPopupMapPickerContext *>(pContext);
+	CMenus *pMenus = pPopupContext->m_pMenus;
+
+	static CListBox s_ListBox;
+	s_ListBox.SetActive(Active);
+	s_ListBox.DoStart(20.0f, pPopupContext->m_vMaps.size(), 1, 3, -1, &View, false);
+
+	int MapIndex = 0;
+	for(auto &Map : pPopupContext->m_vMaps)
+	{
+		MapIndex++;
+		const CListboxItem Item = s_ListBox.DoNextItem(&Map, MapIndex == pPopupContext->m_Selection);
+		if(!Item.m_Visible)
+			continue;
+
+		CUIRect Label, Icon;
+		Item.m_Rect.VSplitLeft(20.0f, &Icon, &Label);
+
+		char aLabelText[IO_MAX_PATH_LENGTH];
+		str_copy(aLabelText, Map.m_aFilename);
+		if(Map.m_IsDirectory)
+			str_append(aLabelText, "/");
+
+		const char *pIconType;
+		if(!Map.m_IsDirectory)
+		{
+			pIconType = FontIcon::MAP;
+		}
+		else
+		{
+			if(!str_comp(Map.m_aFilename, ".."))
+				pIconType = FontIcon::FOLDER_TREE;
+			else
+				pIconType = FontIcon::FOLDER;
+		}
+
+		pMenus->TextRender()->SetFontPreset(EFontPreset::ICON_FONT);
+		pMenus->TextRender()->SetRenderFlags(ETextRenderFlags::TEXT_RENDER_FLAG_ONLY_ADVANCE_WIDTH | ETextRenderFlags::TEXT_RENDER_FLAG_NO_X_BEARING | ETextRenderFlags::TEXT_RENDER_FLAG_NO_Y_BEARING);
+		pMenus->Ui()->DoLabel(&Icon, pIconType, 12.0f, TEXTALIGN_ML);
+		pMenus->TextRender()->SetRenderFlags(0);
+		pMenus->TextRender()->SetFontPreset(EFontPreset::DEFAULT_FONT);
+
+		pMenus->Ui()->DoLabel(&Label, aLabelText, 10.0f, TEXTALIGN_ML);
+	}
+
+	const int NewSelected = s_ListBox.DoEnd();
+	pPopupContext->m_Selection = NewSelected >= 0 ? NewSelected : -1;
+	if(s_ListBox.WasItemSelected() || s_ListBox.WasItemActivated())
+	{
+		const CMapListItem &SelectedItem = pPopupContext->m_vMaps[pPopupContext->m_Selection];
+
+		if(SelectedItem.m_IsDirectory)
+		{
+			if(!str_comp(SelectedItem.m_aFilename, ".."))
+			{
+				dbg_assert(fs_parent_dir(pPopupContext->m_aCurrentMapFolder) == 0, "Parent folder item selected but there is no parent folder");
+			}
+			else
+			{
+				str_append(pPopupContext->m_aCurrentMapFolder, "/");
+				str_append(pPopupContext->m_aCurrentMapFolder, SelectedItem.m_aFilename);
+			}
+			pPopupContext->MapListPopulate();
+		}
+		else
+		{
+			str_format(g_Config.m_ClBackgroundEntities, sizeof(g_Config.m_ClBackgroundEntities), "%s/%s", pPopupContext->m_aCurrentMapFolder, SelectedItem.m_aFilename);
+			pMenus->GameClient()->m_Background.LoadBackground();
+			return CUi::POPUP_CLOSE_CURRENT;
+		}
+	}
+
+	return CUi::POPUP_KEEP_OPEN;
+}
+
+void CMenus::CPopupMapPickerContext::MapListPopulate()
+{
+	m_vMaps.clear();
+	char aTemp[IO_MAX_PATH_LENGTH];
+	str_format(aTemp, sizeof(aTemp), "maps/%s", m_aCurrentMapFolder);
+	m_pMenus->Storage()->ListDirectoryInfo(IStorage::TYPE_ALL, aTemp, MapListFetchCallback, this);
+	std::stable_sort(m_vMaps.begin(), m_vMaps.end(), CompareFilenameAscending);
+	m_Selection = -1;
+}
+
+int CMenus::CPopupMapPickerContext::MapListFetchCallback(const CFsFileInfo *pInfo, int IsDir, int StorageType, void *pUser)
+{
+	CPopupMapPickerContext *pRealUser = (CPopupMapPickerContext *)pUser;
+	if((!IsDir && !str_endswith(pInfo->m_pName, ".map")) || !str_comp(pInfo->m_pName, ".") || (!str_comp(pInfo->m_pName, "..") && (!str_comp(pRealUser->m_aCurrentMapFolder, ""))))
+		return 0;
+
+	CMapListItem Item;
+	str_copy(Item.m_aFilename, pInfo->m_pName);
+	Item.m_IsDirectory = IsDir;
+
+	pRealUser->m_vMaps.emplace_back(Item);
+
+	return 0;
+}
